+++ packages/frontend/src/components/common/Filters/FilterInputs/DefaultFilterInputs.tsx
@@ -1,37 +1,41 @@
 import { NumericInput, TagInput } from '@blueprintjs/core';
 import { Popover2Props } from '@blueprintjs/popover2';
 import {
+    ConditionalRule,
     FilterableField,
     FilterOperator,
-    FilterRule,
     FilterType,
+    isFilterRule,
 } from '@lightdash/common';
 import { isString } from 'lodash-es';
-import React, { FC } from 'react';
+import React from 'react';
 import { useFiltersContext } from '../FiltersProvider';
 import MultiAutoComplete from './AutoComplete/MultiAutoComplete';
 
-export type FilterInputsProps<T extends FilterRule = FilterRule> = {
+export type FilterInputsProps<T extends ConditionalRule> = {
     filterType: FilterType;
     field: FilterableField;
-    filterRule: T;
-    onChange: (value: FilterRule) => void;
+    rule: T;
+    onChange: (value: T) => void;
     popoverProps?: Popover2Props;
     disabled?: boolean;
 };
 
-const DefaultFilterInputs: FC<FilterInputsProps> = ({
+const DefaultFilterInputs = <T extends ConditionalRule>({
     field,
     filterType,
-    filterRule,
+    rule,
     popoverProps,
     disabled,
     onChange,
-}) => {
+}: React.PropsWithChildren<FilterInputsProps<T>>) => {
     const { getField } = useFiltersContext();
-    const suggestions = getField(filterRule)?.suggestions;
-    const filterOperator = filterRule.operator;
-    switch (filterRule.operator) {
+    const suggestions = isFilterRule(rule)
+        ? getField(rule)?.suggestions
+        : undefined;
+
+    const filterOperator = rule.operator;
+    switch (rule.operator) {
         case FilterOperator.NULL:
         case FilterOperator.NOT_NULL:
             return <span style={{ width: '100%' }} />;
@@ -45,12 +49,12 @@ const DefaultFilterInputs: FC<FilterInputsProps> = ({
                     <MultiAutoComplete
                         disabled={disabled}
                         field={field}
-                        values={(filterRule.values || []).filter(isString)}
+                        values={(rule.values || []).filter(isString)}
                         suggestions={suggestions || []}
                         popoverProps={popoverProps}
                         onChange={(values) =>
                             onChange({
-                                ...filterRule,
+                                ...rule,
                                 values,
                             })
                         }
@@ -70,10 +74,10 @@ const DefaultFilterInputs: FC<FilterInputsProps> = ({
                                 : 'text',
                     }}
                     tagProps={{ minimal: true }}
-                    values={filterRule.values || []}
+                    values={rule.values || []}
                     onChange={(values) =>
                         onChange({
-                            ...filterRule,
+                            ...rule,
                             values,
                         })
                     }
@@ -87,7 +91,7 @@ const DefaultFilterInputs: FC<FilterInputsProps> = ({
         case FilterOperator.IN_THE_PAST:
         case FilterOperator.IN_THE_NEXT:
         case FilterOperator.IN_THE_CURRENT:
-            const parsedValue = parseInt(filterRule.values?.[0], 10);
+            const parsedValue = parseInt(rule.values?.[0] as string, 10);
             return (
                 <NumericInput
                     className={disabled ? 'disabled-filter' : ''}
@@ -97,14 +101,14 @@ const DefaultFilterInputs: FC<FilterInputsProps> = ({
                     min={0}
                     onValueChange={(value) =>
                         onChange({
-                            ...filterRule,
+                            ...rule,
                             values: [value],
                         })
                     }
                 />
             );
         default: {
-            const never: never = filterRule.operator;
+            const never: never = rule.operator;
             throw Error(
                 `No form implemented for String filter operator ${filterOperator}`,
             );