+++ packages/frontend/src/hooks/cartesianChartConfig/useCartesianChartConfig.ts
@@ -3,15 +3,21 @@ import {
     CartesianChart,
     CartesianSeriesType,
     ChartType,
+    CompleteCartesianChartLayout,
     EchartsGrid,
     EchartsLegend,
     Explore,
     getSeriesId,
     isCompleteEchartsConfig,
     isCompleteLayout,
+    MarkLineData,
     Series,
 } from '@lightdash/common';
 import { useCallback, useEffect, useMemo, useState } from 'react';
+import {
+    getMarkLineAxis,
+    ReferenceLineField,
+} from '../../components/common/ReferenceLine';
 import {
     getExpectedSeriesMap,
     mergeExistingAndExpectedSeries,
@@ -28,6 +34,63 @@ type Args = {
     >;
     columnOrder: string[];
     explore: Explore | undefined;
+};
+
+const applyReferenceLines = (
+    series: Series[],
+    dirtyLayout: Partial<Partial<CompleteCartesianChartLayout>> | undefined,
+    referenceLines: ReferenceLineField[],
+): Series[] => {
+    let appliedReferenceLines: string[] = []; // Don't apply the same reference line to multiple series
+    return series.map((serie) => {
+        const referenceLinesForSerie = referenceLines.filter(
+            (referenceLine) => {
+                if (referenceLine.fieldId === undefined) return false;
+                if (appliedReferenceLines.includes(referenceLine.fieldId))
+                    return false;
+                return (
+                    referenceLine.fieldId === serie.encode?.xRef.field ||
+                    referenceLine.fieldId === serie.encode?.yRef.field
+                );
+            },
+        );
+
+        if (referenceLinesForSerie.length === 0) return serie;
+        const markLineData: MarkLineData[] = referenceLinesForSerie.map(
+            (line) => {
+                if (line.fieldId === undefined) return line.data;
+                const value = line.data.xAxis || line.data.yAxis;
+                if (value === undefined) return line.data;
+                appliedReferenceLines.push(line.fieldId);
+
+                const axis = getMarkLineAxis(
+                    dirtyLayout?.xField,
+                    dirtyLayout?.flipAxes || false,
+                    line.fieldId,
+                );
+
+                return {
+                    ...line.data,
+                    xAxis: undefined,
+                    yAxis: undefined,
+                    [axis]: value,
+                };
+            },
+        );
+
+        return {
+            ...serie,
+            markLine: {
+                symbol: 'none',
+                lineStyle: {
+                    color: '#000',
+                    width: 3,
+                    type: 'solid',
+                },
+                data: markLineData,
+            },
+        };
+    });
 };
 
 const useCartesianChartConfig = ({
@@ -455,6 +518,42 @@ const useCartesianChartConfig = ({
         setType,
     ]);
 
+    const selectedReferenceLines: ReferenceLineField[] = useMemo(() => {
+        if (dirtyEchartsConfig?.series === undefined) return [];
+        return dirtyEchartsConfig.series.reduce<ReferenceLineField[]>(
+            (acc, serie) => {
+                const data = serie.markLine?.data;
+                if (data !== undefined) {
+                    const referenceLine = data.map((markData) => {
+                        const axis =
+                            markData.xAxis !== undefined
+                                ? dirtyLayout?.flipAxes
+                                    ? serie.encode.yRef
+                                    : serie.encode.xRef
+                                : dirtyLayout?.flipAxes
+                                ? serie.encode.xRef
+                                : serie.encode.yRef;
+                        return {
+                            fieldId: axis.field,
+                            data: {
+                                label: serie.markLine?.label,
+                                lineStyle: serie.markLine?.lineStyle,
+                                ...markData,
+                            },
+                        };
+                    });
+
+                    return [...acc, ...referenceLine];
+                }
+                return acc;
+            },
+            [],
+        );
+    }, [dirtyEchartsConfig?.series, dirtyLayout?.flipAxes]);
+
+    const [referenceLines, setReferenceLines] = useState<ReferenceLineField[]>(
+        selectedReferenceLines,
+    );
     // Generate expected series
     useEffect(() => {
         if (isCompleteLayout(dirtyLayout) && resultsData) {
@@ -479,13 +578,27 @@ const useCartesianChartConfig = ({
                     expectedSeriesMap,
                     existingSeries: prev?.series || [],
                 });
+
+                const seriesWithReferenceLines = applyReferenceLines(
+                    newSeries,
+                    dirtyLayout,
+                    referenceLines,
+                );
+
                 return {
                     ...prev,
-                    series: newSeries,
+                    series: seriesWithReferenceLines,
                 };
             });
         }
-    }, [dirtyLayout, pivotKeys, resultsData, availableDimensions, isStacked]);
+    }, [
+        dirtyLayout,
+        pivotKeys,
+        resultsData,
+        availableDimensions,
+        isStacked,
+        referenceLines,
+    ]);
 
     const validCartesianConfig: CartesianChart | undefined = useMemo(
         () =>
@@ -537,6 +650,8 @@ const useCartesianChartConfig = ({
         setShowGridY,
         setInverseX,
         updateSeries,
+        referenceLines,
+        setReferenceLines,
     };
 };